+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 - Menambahkan Unit Test dan Integration Test
+- Menambahkan utils service untuk payment gateway tripay dan integration test nya
+
 
 ### [2026-09-11]